refactor(core): prep pipeline modules for extraction (Phase 2a U2)
(a) file_policy + indexing_state → julie-core
- crates/julie-core/src/file_policy.rs: all pub(crate)→pub; imports fixed
  (crate::tools::shared → crate::shared, crate::utils::file_utils → crate::file_utils)
- crates/julie-core/src/indexing_state.rs: all pub(crate)→pub; parsed_file_count/
  text_only_file_count unconditionally pub (called from top-crate integration tests —
  cfg(test) in a dep is invisible to dependents)
- Top-crate shims at original paths; IndexedFileState excluded (tests access via
  field, not by name); IndexedFileDisposition re-exported unconditionally

(b) resolve_pending_relationships confirmed handler-free (no code change)

(c) Embedding log-fields → src/embeddings/log_fields.rs
- EmbeddingRuntimeLogFields + build_embedding_runtime_log_fields +
  embedding_telemetry_confidence moved to src/embeddings/log_fields.rs (owns
  its own EmbeddingRuntimeStatus/DeviceInfo types locally)
- src/workspace/mod.rs: definitions removed; cfg(test) re-export for
  build_embedding_runtime_log_fields (test callers use crate::workspace:: path);
  EmbeddingRuntimeLogFields not re-exported (used via return-type inference only)
- src/embeddings/init.rs: upward workspace import replaced with local log_fields import

(d) indexing_core::extraction severed from ManageWorkspaceTool
- process_file_with_parser closure now calls crate::extractors::extract_canonical
  directly instead of ManageWorkspaceTool::extract_symbols_static
- extractor.rs tombstoned (function was the only item, now dead)

Gate: cargo nextest run -p julie-core --no-run + -p julie --no-run: 0 errors, 0 warnings
Tripwires: no_upward_deps manifest + source: 2/2 PASS
