+++ .memories/2026-06-04/072447_ebad.md
@@ -1,0 +1,67 @@
+---
+id: checkpoint_ebad0786
+timestamp: 2026-06-04T07:24:47.999Z
+tags:
+  - julie-rescue
+  - phase2a
+  - crate-split
+  - julie-core
+  - refactor
+git:
+  branch: julie-rescue-phase2a
+  commit: 551634c8
+  files:
+    - .codex/config.toml
+    - .miller/.symbols.db.init.lock
+    - .miller/indexer.lock
+    - .miller/logs/miller-20260531.jsonl
+    - .miller/logs/miller-20260603.jsonl
+    - .miller/logs/miller-35698-20260531.jsonl
+    - .miller/logs/miller-38072-20260531.jsonl
+    - .miller/logs/miller-43360-20260531.jsonl
+    - .miller/symbols.db
+    - .miller/symbols.db.julie-extract.lock
+summary: "U1 complete: P0 leaf relocations → julie-core (Phase 2a)"
+briefId: julie-rescue-de-slop-packaging-keep-the-moat
+type: checkpoint
+evidence:
+  - "cargo nextest run -p julie-core --no-run: PASS"
+  - "cargo build: PASS (1 warning)"
+  - "no_upward_source_references tripwire: PASS"
+  - commit 551634c8
+next: "U2: prep pipeline modules to be movable"
+---
+
+## U1 complete: P0 leaf relocations → julie-core (Phase 2a)
+
+### WHAT
+Moved all shared leaf code down into `crates/julie-core` so nothing above it imports `crate::tools`. Commit `551634c8` on branch `julie-rescue-phase2a`.
+
+### WHY
+Sever the `utils→tools` circular dependency edges that prevented extracting `julie-pipeline` in later sub-PRs.
+
+### HOW
+16 new files added to `crates/julie-core/src/`:
+- `shared.rs` — BLACKLISTED_DIRECTORIES/EXTENSIONS/FILENAMES, NOISE_CALLEE_NAMES, OptimizedResponse
+- `workspace_errors.rs` — WorkspaceResolutionFailure + Kind + workspace_resolution_failure_kind
+- `workspace/{mod,mutation_gate,registry,root_safety,startup_hint}.rs`
+- `external_extract_paths.rs` — ExternalFilePath + normalize_* helpers
+- `health_types.rs` — pure state enums only (compound types stay in top crate due to orphan rule)
+- `serde_lenient.rs`, `token_estimation.rs`, `string_similarity.rs`, `file_utils.rs`, `language.rs`
+- `cross_language_intelligence.rs` (changed `crate::extractors::SymbolKind` → `julie_extractors::SymbolKind`)
+- `mcp_compat.rs`
+
+Top-crate files replaced with `pub use julie_core::...` re-exports (zero churn for callers).
+
+Critical edge cuts:
+- `src/utils/walk.rs`: BLACKLISTED_DIRECTORIES now from `julie_core::shared`
+- `src/utils/paths.rs`: WorkspaceResolution* now from `julie_core::workspace_errors`
+- `src/tools/navigation/resolution.rs`: local duplicate definitions removed, re-exports added
+
+Key fix: `pub(crate)` → `pub` on `root_safety` functions (called from top crate) and `mutation_gate::try_acquire` (called from `src/watcher/observability.rs`). Health compound types kept in top crate to avoid orphan rule violation on `render_report` impl.
+
+### IMPACT
+- `cargo nextest run -p julie-core --no-run` PASS
+- `cargo build` PASS (1 unused-import warning from test-only callers of re-exported names — not an error)
+- `no_upward_source_references` tripwire PASS
+- Unblocks U2/U3/U4 (pipeline module extraction)
