+++ .memories/2026-06-04/073115_c7df.md
@@ -1,0 +1,47 @@
+---
+id: checkpoint_c7df4bd0
+timestamp: 2026-06-04T07:31:15.857Z
+tags:
+  - julie-rescue
+  - phase2a
+  - fix
+  - cfg-test
+git:
+  branch: julie-rescue-phase2a
+  commit: 435fc2f6
+  files:
+    - .codex/config.toml
+    - .miller/.symbols.db.init.lock
+    - .miller/indexer.lock
+    - .miller/logs/miller-20260531.jsonl
+    - .miller/logs/miller-20260603.jsonl
+    - .miller/logs/miller-35698-20260531.jsonl
+    - .miller/logs/miller-38072-20260531.jsonl
+    - .miller/logs/miller-43360-20260531.jsonl
+    - .miller/symbols.db
+    - .miller/symbols.db.julie-extract.lock
+summary: "U1 follow-up: test binary compile fix + warning cleanup"
+briefId: julie-rescue-de-slop-packaging-keep-the-moat
+type: checkpoint
+evidence:
+  - "cargo nextest run -p julie --no-run: PASS, zero errors, zero warnings"
+  - "cargo nextest run -p julie-core --no-run: PASS"
+  - commit 435fc2f6
+---
+
+## U1 follow-up: test binary compile fix + warning cleanup
+
+### WHAT
+Two post-relocation compile fixes committed at `435fc2f6`.
+
+### WHY
+`cargo build` skips `#[cfg(test)]` so the original U1 gate gave a false green. The real gate `cargo nextest run -p julie --no-run` revealed the test binary couldn't compile.
+
+### HOW
+1. **`sanitize_name` visibility**: Was `pub(crate)` in `julie_core::workspace::registry`. Test at `src/tests/tools/workspace/registry.rs` uses glob import `use crate::workspace::registry::*` and calls it directly. Fixed: `pub(crate)` → `pub` in julie-core; added `#[cfg(test)] pub use julie_core::workspace::registry::sanitize_name` in top-crate re-export.
+
+2. **Unused import warning**: `is_sensitive_workspace_root` and `sensitive_root_candidates` in `src/workspace/root_safety.rs` were re-exported unconditionally but only used in test code. Fixed by gating their re-export with `#[cfg(test)]`.
+
+### IMPACT
+- `cargo nextest run -p julie --no-run` — PASS, zero errors, zero warnings
+- `cargo nextest run -p julie-core --no-run` — PASS
