+++ .memories/2026-06-04/074632_6127.md
@@ -1,0 +1,70 @@
+---
+id: checkpoint_6127c6fd
+timestamp: 2026-06-04T07:46:32.687Z
+tags:
+  - julie-rescue
+  - phase2a
+  - U2
+  - crate-split
+git:
+  branch: julie-rescue-phase2a
+  commit: 0b158450
+  files:
+    - .codex/config.toml
+    - .miller/.symbols.db.init.lock
+    - .miller/indexer.lock
+    - .miller/logs/miller-20260531.jsonl
+    - .miller/logs/miller-20260603.jsonl
+    - .miller/logs/miller-35698-20260531.jsonl
+    - .miller/logs/miller-38072-20260531.jsonl
+    - .miller/logs/miller-43360-20260531.jsonl
+    - .miller/symbols.db
+    - .miller/symbols.db.julie-extract.lock
+    - crates/julie-core/src/file_policy.rs
+    - crates/julie-core/src/indexing_state.rs
+    - crates/julie-core/src/lib.rs
+    - src/embeddings/init.rs
+    - src/embeddings/log_fields.rs
+    - src/embeddings/mod.rs
+    - src/indexing_core/extraction.rs
+    - src/tools/workspace/indexing/extractor.rs
+    - src/tools/workspace/indexing/file_policy.rs
+    - src/tools/workspace/indexing/state.rs
+    - src/workspace/mod.rs
+summary: "U2 complete: prep pipeline modules (Phase 2a)"
+briefId: julie-rescue-de-slop-packaging-keep-the-moat
+type: checkpoint
+evidence:
+  - "cargo nextest run -p julie-core --no-run: PASS 0 warnings"
+  - "cargo nextest run -p julie --no-run: PASS 0 warnings"
+  - "no_upward_deps tripwire: 2/2 PASS"
+---
+
+## U2 complete: prep pipeline modules (Phase 2a)
+
+### WHAT
+Four sub-tasks to make pipeline modules extractable into julie-pipeline in U3.
+
+### HOW
+
+**(a) file_policy.rs + indexing state.rs → julie-core**
+- `crates/julie-core/src/file_policy.rs`: changed `crate::tools::shared::` → `crate::shared::`, `crate::utils::file_utils::` → `crate::file_utils::`, all `pub(crate)` → `pub`
+- `crates/julie-core/src/indexing_state.rs`: all `pub(crate)` → `pub`; CRITICAL: `parsed_file_count`/`text_only_file_count` must stay unconditionally `pub` (NOT `#[cfg(test)]`) because they're called from top-crate integration tests — `#[cfg(test)]` in a dep is never compiled into the dependent
+- Top-crate shims at their original paths use `pub use julie_core::...`
+- `IndexedFileState` excluded from re-export (tests access field, not type name); `IndexedFileDisposition` re-exported unconditionally (tests import it by name)
+
+**(b) resolve_pending_relationships confirmed handler-free**
+- No code changes needed; confirmed by reading finalize.rs — only `analyze_batch` takes `&JulieServerHandler`
+
+**(c) embedding log fields → src/embeddings/log_fields.rs**
+- New `src/embeddings/log_fields.rs` with `EmbeddingRuntimeLogFields`, `build_embedding_runtime_log_fields`, `embedding_telemetry_confidence` (all using local `crate::embeddings::` types)
+- `src/workspace/mod.rs`: removed definitions, added `#[cfg(test)] pub(crate) use crate::embeddings::log_fields::build_embedding_runtime_log_fields` (test-only, `EmbeddingRuntimeLogFields` not re-exported since tests use it via return type inference)
+- `src/embeddings/init.rs`: changed import from `crate::workspace::build_embedding_runtime_log_fields` → `crate::embeddings::log_fields::build_embedding_runtime_log_fields`
+
+**(d) extract_symbols_static → direct crate::extractors::extract_canonical**
+- `src/indexing_core/extraction.rs:318` closure: removed `ManageWorkspaceTool::extract_symbols_static` call, replaced with `crate::extractors::extract_canonical`
+- `src/tools/workspace/indexing/extractor.rs`: tombstoned (function was dead code)
+
+### IMPACT
+- Both compile gates: zero errors, zero warnings
+- Both no_upward_deps tripwires: PASS
