+++ crates/julie-core/src/lib.rs
@@ -9,9 +9,11 @@ pub mod cross_language_intelligence;
 pub mod database;
 pub mod embeddings_contract;
 pub mod external_extract_paths;
+pub mod file_policy;
 pub mod file_utils;
 pub mod glob;
 pub mod health_types;
+pub mod indexing_state;
 pub mod language;
 pub mod mcp_compat;
 pub mod paths;
