+++ src/embeddings/init.rs
@@ -13,7 +13,7 @@ use crate::embeddings::{
     resolve_backend_preference, should_disable_for_strict_acceleration,
     strict_acceleration_enabled_from_env_value,
 };
-use crate::workspace::build_embedding_runtime_log_fields;
+use crate::embeddings::log_fields::build_embedding_runtime_log_fields;
 
 /// Create an embedding provider by reading environment variables and resolving
 /// the backend preference. Returns the provider (if successful) and runtime
